requirements.txt
fastapi
uvicorn[standard]
langdetect
deep-translator
openai
python-dotenv
pdfplumber
SpeechRecognition
edge-tts
python-multipart
pydub
slowapi
aiofiles
gunicorn
